Changes completeOrder.test.js to completeOrderV.test.js
