--- test/completeOrder.test.js
@@ -1,29 +1,0 @@
-"use strict";
-
-const { assert: { deepEqual, equal } } = require('chai'),
-      { createPrompt } = require('../app/views/completeOrderV');
-
-describe('completeOrder module', () => {
-  describe('completeOrder view', () => {
-    describe('createPrompt()', () => {
-      const checkout = {
-            description: 'Your order total is $180. Ready to purchase?\n(Y/N)'
-          },
-          payment = {
-            description: 'Choose a payment option:\n1. AmEx\n2. Visa',
-            pattern: new RegExp('^[1-2]$'),
-          };
-      const [a, b] = createPrompt(180, ['AmEx', 'Visa']);
-
-      it('should create a description with the order total included', () => {
-	       equal(a.description, checkout.description);
-      });
-      it('should create a description with the customer\'s payment options', () => {
-	       equal(b.description, payment.description);
-      });
-      it('should create a RegEx for payment that only allows the number of payment options a customer has', () => {
-	       deepEqual(b.pattern, payment.pattern);
-      });
-    });
-  });
-});
